[AMTH] Refactor utils
utils.h contains all the general utils. All methods are under a different
namespace.

mth_utils.h contains all utils related MTH. All methods are under the namespace
multi_theory_horn.
