+++ src/Int2BvPreprocessor.h
@@ -5,7 +5,7 @@
 //------------------------------------------------------------------------------
 
 #pragma once
-#include "utils.h"
+#include "mth_utils.h"
 #include "InstCombiner.h"
 #include <z3++.h>
 #include <unordered_set>
